+++ hooks/scrollToTop.tsx
@@ -1,7 +1,7 @@
 
 export const useScrollToTop = () => {
 
-  const scrollToTop = (myRef: React.RefObject<HTMLDivElement>) => {
+  const scrollToTop = (myRef: React.RefObject<HTMLDivElement | null>) => {
     if (myRef.current) {
       myRef.current.scrollIntoView({ behavior: 'smooth' });
       myRef.current.focus();